Hardcode valid time range for reproducible builds
Previously each build of `nerves_time` resulted in a different binary
due to the use of a build timestamp in sane_time.ex. This commit fixes
that by hardcoding a valid time window. This also adds the ability for
applications to override the time window should they want something more
recent.