Bulletproof make_sane/1
`make_sane/1` is intended to make any time be reasonable. If you give it
something that's not a time, I made it crash in a refactor. It turns out
that this behavior was actually used! The code that caused this is fixed
now, but the `nerves_time` crash bothered me enough that it seems safest
to make `make_sane/1` robust to nonsense and carry on even though it
isn't the Erlang way.