Rename initial RTC time adjustment functions; fix
This makes the functions slighty more clear without doing a full
refactor. It also removes the time set on terminate which 1) wasn't
tested, and 2) probably isn't what we want anyway. For context, the
original thought was that if `nerves_time` is going down, it should
update the RTC to whatever latest and greatest thing that NTP has to
offer. My current thinking is that the RTC can't possibly be so far off
to risk crashing or messing it up on terminate.