Add terminate callback on RTC for final ops
This is desirable for the FileTime implementation so that it can touch
the file one final time before the restart/reboot/whatever. This
reduces the lost time especially in situations where NTP isn't working
so well.

This undoes the functionality that that commit
5da6fbd72dd1465500ce312bbd446a8a671061bd removed, but in a way that RTC
modules can ignore if they want. (Which is probably what they want since
they already know the time and it's best to avoid I2C or other
transactions on a terminate)